+++ src/engine.rs
@@ -22,6 +22,7 @@ pub mod issue_body;
 pub mod issue_cache;
 pub mod issue_map;
 pub mod lease;
+pub mod milestone_cache;
 pub mod prompt;
 pub mod provenance;
 pub mod refs;
